refactor: stop アクションを AlarmMembershipsController に統一する
